fix: Statvfs.Spare field type ulong[] → int[] to match Linux ABI (D4)
The Statvfs struct's Spare field was declared as ulong[] (48 bytes), but the
real Linux statvfs struct has int __f_spare[6] (24 bytes). This ABI mismatch
caused 24 bytes of over-allocation. Changed field type to int[] to match the
correct binary layout.

D5 (COM reference leak) verified already fixed in WindowsSystemMetricsProvider
at line ~967 with full Dispose() implementation.

+++ src/NexusMonitor.Platform.Linux/LinuxSystemMetricsProvider.cs
@@ -28,7 +28,7 @@ public sealed class LinuxSystemMetricsProvider : ISystemMetricsProvider, IDispos
         public ulong f_flag;
         public ulong f_namemax;
         [MarshalAs(UnmanagedType.ByValArray, SizeConst = 6)]
-        public ulong[] Spare;
+        public int[] Spare;
     }
 
     // ── Cached static data ─────────────────────────────────────────────────────
